+++ decomp/ST.exe/functions/00491070/decomp.c
@@ -40,7 +40,7 @@ void __thiscall STBoatC::ReleaseLoad(STBoatC *this,uint param_1)
       this->field_0076 = 1;
       return;
     }
-    iVar2 = this->vfunc_AC(param_1);
+    iVar2 = this->vfunc_AC((short)param_1);
     if (iVar2 == 1) {
       this->field_05C0 = 2;
     }
