+++ decomp/ST.exe/functions/00586240/decomp.c
@@ -37,7 +37,7 @@ FUN_00586240(int param_1,int param_2,int param_3,undefined4 param_4,undefined4 p
   local_30[2] = 0;
   local_30[3] = 1;
   local_8 = param_7;
-  g_playSystem_00802A38->vfunc_08(0x131,0,&param_5,local_30,0);
+  g_playSystem_00802A38->vfunc_08(0x131,0,(short)&param_5,(short)local_30,0);
   return param_5;
 }
