+++ decomp/ST.exe/functions/0058BAB0/decomp.c
@@ -26,7 +26,7 @@ FUN_0058bab0(undefined4 param_1,undefined4 param_2,undefined4 param_3,undefined4
   local_2c[2] = 0;
   local_2c[3] = 1;
   local_10 = param_3;
-  g_playSystem_00802A38->vfunc_08(0x14d,0,&param_4,local_2c,0);
+  g_playSystem_00802A38->vfunc_08(0x14d,0,(short)&param_4,(short)local_2c,0);
   return param_4;
 }
