+++ decomp/ST.exe/functions/005E9970/decomp.c
@@ -17,18 +17,18 @@ void __thiscall WaitTy::sub_005E9970(WaitTy *this)
   if (g_startSystem_0081176C->field_0389 != 0) {
     this->field_0025 = g_startSystem_0081176C->field_0389;
     *(undefined4 *)&this->field_0x31 = 1;
-    (*this->field_000C->vtable->vfunc_18)(&this->field_0x1d);
+    (*this->field_000C->vtable->vfunc_18)((short)&this->field_0x1d);
   }
   if (g_startSystem_0081176C->field_0550 != 0) {
     this->field_0025 = g_startSystem_0081176C->field_0550;
     *(undefined4 *)&this->field_0x31 = 1;
-    (*this->field_000C->vtable->vfunc_18)(&this->field_0x1d);
+    (*this->field_000C->vtable->vfunc_18)((short)&this->field_0x1d);
   }
   if (g_startSystem_0081176C->field_054C != 0) {
     this->field_0025 = g_startSystem_0081176C->field_054C;
     this->field_002D = 0x29;
     *(undefined4 *)&this->field_0x31 = 9;
-    (*this->field_000C->vtable->vfunc_18)(&this->field_0x1d);
+    (*this->field_000C->vtable->vfunc_18)((short)&this->field_0x1d);
   }
   return;
 }
