+++ decomp/ST.exe/functions/0061F5C0/decomp.c
@@ -46,7 +46,7 @@ FUN_0061f5c0(undefined4 param_1,ushort param_2,ushort param_3,ushort param_4,sho
   local_30[0] = 0x168;
   local_30[1] = 0xff;
   local_16 = param_5;
-  g_playSystem_00802A38->vfunc_08(0x14c,0,0,local_30,0);
+  g_playSystem_00802A38->vfunc_08(0x14c,0,0,(short)local_30,0);
   return;
 }
