+++ decomp/ST.exe/functions/0062B770/decomp.c
@@ -46,7 +46,7 @@ FUN_0062b770(undefined4 param_1,undefined4 param_2,undefined4 param_3,undefined4
   local_2c = 0;
   local_30 = 0;
   local_c = param_7;
-  g_playSystem_00802A38->vfunc_08(0x124,0,0,local_44,0);
+  g_playSystem_00802A38->vfunc_08(0x124,0,0,(short)local_44,0);
   return;
 }
