+++ decomp/ST.exe/functions/00660540/decomp.c
@@ -46,7 +46,7 @@ FUN_00660540(void *this,undefined4 param_1,short param_2,short param_3,short par
   {
     this_00 = thunk_FUN_0042b760(*(char *)((int)this + 0x24),*(ushort *)((int)this + 0x7d));
     if (this_00 != nullptr) {
-      this_00->sub_00498D20(9,&local_24);
+      this_00->sub_00498D20(9,(short)&local_24);
       return 0;
     }
   }
