+++ decomp/ST.exe/functions/00660700/decomp.c
@@ -80,7 +80,7 @@ FUN_00660700(void *this,short param_1,short param_2,short param_3,int param_4,in
   if (((*(ushort *)((int)this + 0x7d) != 0xfffe) && (g_allPlayers_007FA174 != nullptr))
      && (this_01 = thunk_FUN_0042b760(*(char *)((int)this + 0x24),*(ushort *)((int)this + 0x7d)),
         this_01 != nullptr)) {
-    this_01->sub_00498D20(0x10,&local_14);
+    this_01->sub_00498D20(0x10,(short)&local_14);
     return 0;
   }
   return 0xffffffff;
