+++ decomp/ST.exe/functions/00660A40/decomp.c
@@ -19,7 +19,7 @@ undefined4 __thiscall FUN_00660a40(void *this,short param_1,short param_2,short 
          (g_allPlayers_007FA174 != nullptr)) {
         this_00 = thunk_FUN_0042b760(*(char *)((int)this + 0x24),*(ushort *)((int)this + 0x7d));
         if (this_00 != nullptr) {
-          this_00->sub_00498D20(0xb,&local_8);
+          this_00->sub_00498D20(0xb,(short)&local_8);
           return 0;
         }
       }
