+++ decomp/ST.exe/functions/00660B50/decomp.c
@@ -27,7 +27,7 @@ FUN_00660b50(AnonShape_00660620_6BCED4D7 *param_1,int param_2,short param_3,shor
     if ((param_1->field_007D != 0xfffe) && (g_allPlayers_007FA174 != nullptr)) {
       pSVar1 = thunk_FUN_0042b760(param_1->field_0024,param_1->field_007D);
       if (pSVar1 != nullptr) {
-        pSVar1->sub_00498D20(0xc,&local_c);
+        pSVar1->sub_00498D20(0xc,(short)&local_c);
         return 0;
       }
     }
@@ -46,7 +46,7 @@ FUN_00660b50(AnonShape_00660620_6BCED4D7 *param_1,int param_2,short param_3,shor
           local_c = param_1;
           pSVar1 = thunk_FUN_0042b760(param_1->field_0024,param_1->field_007D);
           if (pSVar1 != nullptr) {
-            pSVar1->sub_00498D20(0x12,&local_14);
+            pSVar1->sub_00498D20(0x12,(short)&local_14);
             return 0;
           }
         }
